python_service/requirements.txt
pika
openai
psycopg2-binary
markdown
python-dotenv